Modify Sluggable trait to use default locale
