+++ src/Model/Sluggable/SluggableMethodsTrait.php
@@ -4,9 +4,6 @@ declare(strict_types=1);
 
 namespace Knp\DoctrineBehaviors\Model\Sluggable;
 
-use Knp\DoctrineBehaviors\Exception\SluggableException;
-use Symfony\Component\String\Slugger\AsciiSlugger;
-
 trait SluggableMethodsTrait
 {
     public function setSlug(string $slug): void
@@ -19,83 +16,18 @@ trait SluggableMethodsTrait
         return $this->slug;
     }
 
-    /**
-     * Generates and sets the entity's slug. Called prePersist and preUpdate
-     */
-    public function generateSlug(): void
-    {
-        if ($this->slug !== null && $this->shouldRegenerateSlugOnUpdate() === false) {
-            return;
-        }
-
-        $values = [];
-        foreach ($this->getSluggableFields() as $sluggableField) {
-            $values[] = $this->resolveFieldValue($sluggableField);
-        }
-
-        $this->slug = $this->generateSlugValue($values);
-    }
-
     public function shouldGenerateUniqueSlugs(): bool
     {
         return false;
     }
 
-    private function getSlugDelimiter(): string
+    public function getSlugDelimiter(): string
     {
         return '-';
     }
 
-    private function shouldRegenerateSlugOnUpdate(): bool
+    public function shouldRegenerateSlugOnUpdate(): bool
     {
         return true;
-    }
-
-    /**
-     * @return mixed|string
-     */
-    private function generateSlugValue($values)
-    {
-        $usableValues = [];
-        foreach ($values as $fieldValue) {
-            if (! empty($fieldValue)) {
-                $usableValues[] = $fieldValue;
-            }
-        }
-
-        $this->ensureAtLeastOneUsableValue($values, $usableValues);
-
-        // generate the slug itself
-        $sluggableText = implode(' ', $usableValues);
-
-        $unicodeString = (new AsciiSlugger())->slug($sluggableText, $this->getSlugDelimiter());
-
-        return strtolower($unicodeString->toString());
-    }
-
-    private function ensureAtLeastOneUsableValue(array $values, array $usableValues): void
-    {
-        if (count($usableValues) >= 1) {
-            return;
-        }
-
-        throw new SluggableException(sprintf(
-            'Sluggable expects to have at least one non-empty field from the following: ["%s"]',
-            implode('", "', array_keys($values))
-        ));
-    }
-
-    private function resolveFieldValue(string $field)
-    {
-        if (property_exists($this, $field)) {
-            return $this->{$field};
-        }
-
-        $methodName = 'get' . ucfirst($field);
-        if (method_exists($this, $methodName)) {
-            return $this->{$methodName}();
-        }
-
-        return null;
     }
 }
